requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.3.1+cu121
accelerate==0.32.1
ipadic
mecab-python3
rank-bm25==0.2.2
rhoknp==1.7.0
safetensors==0.4.3
scikit-learn
scipy
sentencepiece
tqdm
transformers==4.42.4
